Application: process help options immediately
Previously, help output would either be displayed in an existing help
terminal, or just not appear at all. Neither of these are what we want;
we want help to be displayed wherever the program was invoked from.

So instead, we look for the help args in main, and if present we don't
even try to get running terminal instances or new windows involved,
instead just showing the help directly.