split command line logic between local and primary instances
local_command_line() and handle_local_options() are responsible for the
command line parsing and adjusting the options for usage in command_line()
that handles the actual program logic in the primary instance.

this simplifies the primary instance work, and keep command line specific
behavior to be parsed only from command line invocation.