Application: properly unexport the dbus interface on exit
we where depending on the binary exiting to unexport the interface, however,
the test suite launch a instance per test, and they where failing because of
the interface being already registred, and ending in false positives.

the suite where ammended to check for the application exit status too and
fail if not 0.