Enhance unsafe detection (#800)
Not the cleanest. Can always define "more" dangerous behaviours, depends how much we want to get in the weeds.
I need to learn more vala

contains:
Multiple commands: as oneliners chained with ";", "||", "|", "&&", "&"

Admin commands: run0, su, pkexec, who serve similar purpose to sudo and doas. "su -" can do a lot too
Add a space, to avoid false flag if say, it is a word like "visual" that has "su" in it. A command needs a space after it.

Co-authored-by: Jeremy Wootten <jeremywootten@gmail.com>